Overflow, Finalize, Elemental
* int to long for local number of scenarios
* MPI_Finalize is called in StructJuMP not in PIPS
* Elemental call fixed

+++ PIPS-NLP/Core/NlpStoch/sData.C
@@ -210,7 +210,7 @@ sData::~sData()
     delete children[it];
 }
 
-int sData::getLocalnx()
+long sData::getLocalnx()
 {
   StochSymMatrix& Qst = dynamic_cast<StochSymMatrix&>(*H);
   return Qst.diag->size();
